fixing warning, refs #3
git-svn-id: file:///home/behr_mi/git/sumo_synched/trunk@22151 afbd958f-9f77-42d5-a016-97a22340ccf4

+++ sumo/src/netbuild/NBTrafficLightLogicCont.cpp
@@ -293,7 +293,7 @@ NBTrafficLightLogicCont::setTLControllingInformation(const NBEdgeCont& ec, const
             NBOwnTLDef dummy(n->getID(), n, 0, TLTYPE_STATIC);
             dummy.setParticipantsInformation();
             dummy.setTLControllingInformation();
-            n->setCrossingTLIndices(dummy.getID(), dummy.getControlledLinks().size());
+            n->setCrossingTLIndices(dummy.getID(), (int)dummy.getControlledLinks().size());
             n->removeTrafficLight(&dummy);
         }
     }
